improve useStaticSWR
